+++ src/main/java/com/example/concertreservationsystem/infrastructure/config/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.concertreservationsystem.infrastructure.config;
+package com.example.concertreservationsystem.infrastructure.config.swagger;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
